Add `VideoDurationFormatter`
