+++ Sources/StreamChatUI/CommonViews/Attachments/AttachmentViews/VideoAttachmentComposerPreview.swift
@@ -102,8 +102,8 @@ open class VideoAttachmentComposerPreview: _View, ThemeProvider {
                     self?.previewImageView.image = nil
                 }
             }
-            videoDurationLabel.text = appearance.formatters.videoDuration.string(
-                from: components.videoLoader.videoAsset(at: url).duration.seconds
+            videoDurationLabel.text = appearance.formatters.videoDuration.format(
+                components.videoLoader.videoAsset(at: url).duration.seconds
             )
         }
     }
