+++ Sources/StreamChatUI/Gallery/VideoPlaybackControlView.swift
@@ -103,7 +103,7 @@ open class VideoPlaybackControlView: _View, ThemeProvider {
         .withoutAutoresizingMaskConstraints
 
     /// A formatter to convert video duration to textual representation.
-    open lazy var videoDurationFormatter = appearance.formatters.videoDuration
+    open lazy var videoDurationFormatter: VideoDurationFormatter = appearance.formatters.videoDuration
     
     override open func setUp() {
         super.setUp()
@@ -145,16 +145,16 @@ open class VideoPlaybackControlView: _View, ThemeProvider {
         super.setUpAppearance()
         
         playPauseButton.setTitleColor(.black, for: .normal)
-        timestampLabel.text = videoDurationFormatter.string(from: 0)
-        durationLabel.text = videoDurationFormatter.string(from: 0)
+        timestampLabel.text = videoDurationFormatter.format(0)
+        durationLabel.text = videoDurationFormatter.format(0)
     }
     
     override open func updateContent() {
         super.updateContent()
         
         timeSlider.value = .init(content.playingProgress)
-        timestampLabel.text = videoDurationFormatter.string(from: content.currentTime)
-        durationLabel.text = videoDurationFormatter.string(from: content.videoDuration)
+        timestampLabel.text = videoDurationFormatter.format(content.currentTime)
+        durationLabel.text = videoDurationFormatter.format(content.videoDuration)
                 
         switch content.videoState {
         case .playing:
